src/requirements.txt
# Dependências Python para o sistema de agentes de IA
# Usado principalmente pelo Agente Key para geração de conteúdo LLM

# Dependências básicas
python-dotenv>=1.0.0  # Carregamento de variáveis de ambiente

# APIs externas (opcional - descomente se for usar)
# openai>=1.0.0  # Para integração com OpenAI
# anthropic>=0.7.0  # Para integração com Anthropic Claude

# Processamento de dados (opcional - descomente se necessário)
# numpy>=1.21.0
# pandas>=1.3.0

# Machine Learning (opcional - descomente se for usar modelos locais)
# torch>=1.9.0
# transformers>=4.6.1
# tensorflow>=2.5.0
# scikit-learn>=0.24.2

# HTTP requests (opcional - se necessário para scripts Python)
# requests>=2.25.1